+++ Packages/com.vareiko.foundation/Runtime/Push/FoundationPushNotificationInstaller.cs
@@ -1,0 +1,42 @@
+using Zenject;
+
+namespace Vareiko.Foundation.Push
+{
+    public static class FoundationPushNotificationInstaller
+    {
+        public static void Install(DiContainer container, PushNotificationConfig config = null)
+        {
+            if (container.HasBinding<IPushNotificationService>())
+            {
+                return;
+            }
+
+            if (!container.HasBinding<SignalBus>())
+            {
+                SignalBusInstaller.Install(container);
+            }
+
+            container.DeclareSignal<PushInitializedSignal>();
+            container.DeclareSignal<PushPermissionChangedSignal>();
+            container.DeclareSignal<PushTokenUpdatedSignal>();
+            container.DeclareSignal<PushTopicSubscribedSignal>();
+            container.DeclareSignal<PushTopicSubscriptionFailedSignal>();
+            container.DeclareSignal<PushTopicUnsubscribedSignal>();
+            container.DeclareSignal<PushTopicUnsubscriptionFailedSignal>();
+
+            if (config != null)
+            {
+                container.BindInstance(config).IfNotBound();
+            }
+
+            if (config != null && config.Provider == PushNotificationProviderType.Simulated)
+            {
+                container.BindInterfacesAndSelfTo<SimulatedPushNotificationService>().AsSingle().NonLazy();
+            }
+            else
+            {
+                container.Bind<IPushNotificationService>().To<NullPushNotificationService>().AsSingle();
+            }
+        }
+    }
+}